Add first crossover (uniform) function. Tested.

+++ src/randomgenerator.cpp
@@ -10,6 +10,6 @@ namespace RandomGenerator
 
     double get_random_real_between(const double min, const double max)
     {
-        return min + (rand()/RAND_MAX) * (max - min);
+        return min + (rand()/(double)RAND_MAX) * (max - min);
     }
 }
